requirements.txt
Flask==2.3.3
Pillow==10.0.0
matplotlib==3.7.2
numpy==1.24.3
requests==2.31.0
gunicorn==21.2.0